worker/requirements.txt
psycopg2-binary
redis
python-dotenv
pandas